please ignore

+++ commands/Utilities/say.js
@@ -8,15 +8,16 @@ module.exports = class extends Command {
             cooldown: 5,
             permissionLevel: 4,
             requiredPermissions: ["MANAGE_MESSAGES", "USE_EXTERNAL_EMOJIS"],
-            description: msg => msg.language.get("COMMAND_SAY_DESCRIPTION"),
-            usage: "<message:string> [...]",
+            description: language => language.get("COMMAND_SAY_DESCRIPTION"),
+            usage: "<message:string>",
             extendedHelp: "No extended help available."
         });
     }
 
-    async run(msg, [...message]) {
+    async run(msg, [message]) {
+        if (message.length < 1) return msg.sendMessage("<:penguError:435712890884849664> Message length can't be less than one character.");
         msg.delete();
-        return msg.sendMessage(message.join(" "));
+        return msg.sendMessage(message);
     }
 
 };
